+++ src/zigbee/examples/zigbee-nwk-routing.cc
@@ -278,7 +278,7 @@ main(int argc, char* argv[])
     // Configure the Zigbee stack and use only the NWK layer.
 
     ZigbeeHelper zigbeeHelper;
-    zigbeeHelper.SetNwkLayerOnly();
+    zigbeeHelper.SetLayers(StackLayers::ONLY_NWK);
     ZigbeeStackContainer zigbeeStackContainer = zigbeeHelper.Install(lrwpanDevices);
 
     Ptr<ZigbeeStack> zstack0 = zigbeeStackContainer.Get(0)->GetObject<ZigbeeStack>();