+++ RELEASE_NOTES.md
@@ -38,6 +38,7 @@ been tested on Linux. As of this release, the latest known version to work with 
 
 - (network) IANA protocol and link types are now centralized in network module headers.
 - Added support for `nlohmann/json`, a header-only C++ third-party library for JSON parsing and serialization.
+- (zigbee) !2964 Added basic support for Zigbee Device Object (ZDO)
 
 ### Bugs fixed